Editing GeoJSON Offline in QGIS Test
Can't edit GeoJSON natively in QGIS offline:

To edit in QGIS offline:
1) Clone Repo/Branch locally
2) Open local GeoJSON in QGIS
3) Save Copy of file as Shapefile
4) Edit shapefile (attributes or geometry) - Save edits
5) Save newly edited SHP to GeoJSON file with same name as orginal
6)Recommit
7)Pull Request

Used the NGA.geojson file from the GADM REPO (tried to use the test
branch-not master. Not sure if that worked)
After cloning it locally I opened it in QGIS.
Then I saved a copy as a shp to make edits
Resaved the shp as a GeoJSON with the exact same file name as the
original geoJSON.
Then do a Pull Request to have it pulled back into the online Repo.